Decouple/Move URL & Views code to respective apps
With us now having separate apps for Employee related functions & Laptop
related Functions, the URLs & Views should be moved to their respective
apps (currently all urls & views in Hardware app) to keep code logically
placed in project.

Common urls & views (login, dashboard etc) moved to project directory for
the above reason. These urls & views don't belong to any one specific app
and thus the project dir would be the best place for the same.

The URLs have been updated for this change, all 'Employee' urls are prefixed
with 'employee/' and hardware urls with 'hardware'. This keeps the URls
clean and realated to the respective app views.
