+++ minierp/urls.py
@@ -15,12 +15,21 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-
 from django.conf.urls.static import static
 from django.conf import settings
+from . import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('hardware.urls')),
+
+    #General Views
+    path('login/', views.loginPage, name='login'),
+    path('logout/', views.logoutPage, name='logout'),
+    path('', views.home, name='home'),
+
+    #App Views
+    path('employee/', include('employee.urls')),
+    path('hardware/', include('hardware.urls')),
     path('api/', include('api.urls')),
+    
 ] + static(settings.MEDIA_URL, document_root = settings.MEDIA_ROOT) + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
