+++ minierp/views.py
@@ -1,0 +1,32 @@
+from django.shortcuts import render
+from django.contrib.auth import authenticate, login, logout
+from django.shortcuts import render, redirect
+from hardware.decorators import unauthenticated_user, allowed_users
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+
+@unauthenticated_user
+def loginPage(request):
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.info(request, "username OR password is Incorrect.")
+
+    context = {}
+    return render(request, 'login.html', context)
+
+def logoutPage(request):
+    logout(request)
+    return redirect('login')
+
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['admin'])
+def home(request):
+    return render(request, 'dashboard.html')
